More Code Cleaning
I set /names:as-is as compiler option in intel/fortran and cleaned up a lot of inconsistencies. It's just for aesthetics. I reset to default (no /names option) for regular operation.
I also moved the UA code into a separate directory to make it more convenient and clear to eventual users.